backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
python-dotenv
supabase
pypdf
langchain-community
tiktoken
unstructured
lxml
pypandoc
pyjwt
langchain-google-genai
python-multipart
google-generativeai
pymupdf
